requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudscraper==1.2.64
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.6
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.13.1
filelock==3.7.1
flower==1.1.0
greenlet==1.1.2
gunicorn==20.1.0
humanize==4.2.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.4
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.1
packaging==21.3
parsel==1.6.0
Pillow==9.2.0
playwright==1.24.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyee==8.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyTelegramBotAPI==4.6.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
queuelib==1.6.2
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
Scrapy==2.6.2
scrapy-playwright==0.0.20
scrapyrt==0.13.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.1
tornado==6.2
Twisted==22.4.0
typing-extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.1
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.4.0
